Added apply2 to general + triangular and diagonal; fixed triagonal.apply2

+++ src/matrix/ops/gehe.zig
@@ -34,9 +34,9 @@ pub fn apply2(
     if (comptime types.orderOf(@TypeOf(x)) == .col_major) { // orderOf(result) == orderOf(x)
         if (comptime types.orderOf(@TypeOf(y)) == .col_major) {
             if (comptime types.uploOf(@TypeOf(y)) == .upper) { // c c cu
-                var j: usize = 0;
+                var j: u32 = 0;
                 while (j < x.cols) : (j += 1) {
-                    var i: usize = 0;
+                    var i: u32 = 0;
                     while (i < j) : (i += 1) {
                         if (comptime opinfo.@"fn".params.len == 2) {
                             result.data[i + j * result.ld] = op(x.data[i + j * x.ld], y.data[i + j * y.ld]);
@@ -61,7 +61,7 @@ pub fn apply2(
                     }
                 }
             } else { // c c cl
-                var j: usize = 0;
+                var j: u32 = 0;
                 while (j < x.cols) : (j += 1) {
                     if (comptime opinfo.@"fn".params.len == 2) {
                         result.data[j + j * result.ld] = op(x.data[j + j * x.ld], y.data[j + j * y.ld]);
@@ -69,7 +69,7 @@ pub fn apply2(
                         result.data[j + j * result.ld] = try op(x.data[j + j * x.ld], y.data[j + j * y.ld], ctx);
                     }
 
-                    var i: usize = j + 1;
+                    var i: u32 = j + 1;
                     while (i < x.rows) : (i += 1) {
                         if (comptime opinfo.@"fn".params.len == 2) {
                             result.data[i + j * result.ld] = op(x.data[i + j * x.ld], y.data[i + j * y.ld]);
@@ -90,7 +90,7 @@ pub fn apply2(
             }
         } else {
             if (comptime types.uploOf(@TypeOf(y)) == .upper) { // c c ru
-                var j: usize = 0;
+                var j: u32 = 0;
                 while (j < x.cols) : (j += 1) {
                     if (comptime opinfo.@"fn".params.len == 2) {
                         result.data[j + j * result.ld] = op(x.data[j + j * x.ld], y.data[j * y.ld + j]);
@@ -98,7 +98,7 @@ pub fn apply2(
                         result.data[j + j * result.ld] = try op(x.data[j + j * x.ld], y.data[j * y.ld + j], ctx);
                     }
 
-                    var i: usize = j + 1;
+                    var i: u32 = j + 1;
                     while (i < x.rows) : (i += 1) {
                         if (comptime opinfo.@"fn".params.len == 2) {
                             result.data[i + j * result.ld] = op(
@@ -117,9 +117,9 @@ pub fn apply2(
                     }
                 }
             } else { // c c rl
-                var j: usize = 0;
+                var j: u32 = 0;
                 while (j < x.cols) : (j += 1) {
-                    var i: usize = 0;
+                    var i: u32 = 0;
                     while (i < j) : (i += 1) {
                         if (comptime opinfo.@"fn".params.len == 2) {
                             result.data[i + j * result.ld] = op(
@@ -148,9 +148,9 @@ pub fn apply2(
     } else {
         if (comptime types.orderOf(@TypeOf(y)) == .col_major) {
             if (comptime types.uploOf(@TypeOf(y)) == .upper) { // r r cu
-                var i: usize = 0;
+                var i: u32 = 0;
                 while (i < x.rows) : (i += 1) {
-                    var j: usize = 0;
+                    var j: u32 = 0;
                     while (j < i) : (j += 1) {
                         if (comptime opinfo.@"fn".params.len == 2) {
                             result.data[i * result.ld + j] = op(
@@ -175,7 +175,7 @@ pub fn apply2(
                     }
                 }
             } else { // r r cl
-                var i: usize = 0;
+                var i: u32 = 0;
                 while (i < x.rows) : (i += 1) {
                     if (comptime opinfo.@"fn".params.len == 2) {
                         result.data[i * result.ld + i] = op(x.data[i * x.ld + i], y.data[i + i * y.ld]);
@@ -183,7 +183,7 @@ pub fn apply2(
                         result.data[i * result.ld + i] = try op(x.data[i * x.ld + i], y.data[i + i * y.ld], ctx);
                     }
 
-                    var j: usize = i + 1;
+                    var j: u32 = i + 1;
                     while (j < x.cols) : (j += 1) {
                         if (comptime opinfo.@"fn".params.len == 2) {
                             result.data[i * result.ld + j] = op(
@@ -204,7 +204,7 @@ pub fn apply2(
             }
         } else {
             if (comptime types.uploOf(@TypeOf(y)) == .upper) { // r r ru
-                var i: usize = 0;
+                var i: u32 = 0;
                 while (i < x.rows) : (i += 1) {
                     if (comptime opinfo.@"fn".params.len == 2) {
                         result.data[i * result.ld + i] = op(x.data[i * x.ld + i], y.data[i * y.ld + i]);
@@ -212,7 +212,7 @@ pub fn apply2(
                         result.data[i * result.ld + i] = try op(x.data[i * x.ld + i], y.data[i * y.ld + i], ctx);
                     }
 
-                    var j: usize = i + 1;
+                    var j: u32 = i + 1;
                     while (j < x.cols) : (j += 1) {
                         if (comptime opinfo.@"fn".params.len == 2) {
                             result.data[i * result.ld + j] = op(x.data[i * x.ld + j], y.data[i * y.ld + j]);
@@ -231,9 +231,9 @@ pub fn apply2(
                     }
                 }
             } else { // r r rl
-                var i: usize = 0;
+                var i: u32 = 0;
                 while (i < x.rows) : (i += 1) {
-                    var j: usize = 0;
+                    var j: u32 = 0;
                     while (j < i) : (j += 1) {
                         if (comptime opinfo.@"fn".params.len == 2) {
                             result.data[i * result.ld + j] = op(x.data[i * x.ld + j], y.data[i * y.ld + j]);
